Beta 0.5.0
User encryption mechanics! Now, as long as a client is logged in as a user, all data transmitted between the client and server is encrypted with that user's password. This combined with SSL allows for 2 layers of encryption! It's important to know that the user's password is transmitted in plaintext in the beginning. So be sure to use SSL if you want to log in.